feat: add disciplines view and new workflow
