+++ web/frontend/src/components/layout/project/index.ts
@@ -1,2 +1,3 @@
 export { default as ProjectView } from "./project-view";
 export { default as CollaboratorsView } from "./collaborator-view";
+export { default as DisciplinesView } from "./disciplines-view";
